# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * ScenePulse API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type VenueCategory = typeof VenueCategory[keyof typeof VenueCategory];


export const VenueCategory = {
  bar: 'bar',
  restaurant: 'restaurant',
  retail: 'retail',
  cafe: 'cafe',
  experience: 'experience',
} as const;

export type VenueCrowdLevel = typeof VenueCrowdLevel[keyof typeof VenueCrowdLevel];


export const VenueCrowdLevel = {
  open: 'open',
  lively: 'lively',
  packed: 'packed',
} as const;

export type VenueLineTrend = typeof VenueLineTrend[keyof typeof VenueLineTrend];


export const VenueLineTrend = {
  rising: 'rising',
  falling: 'falling',
  steady: 'steady',
} as const;

export type VenueNoiseLevel = typeof VenueNoiseLevel[keyof typeof VenueNoiseLevel];


export const VenueNoiseLevel = {
  quiet: 'quiet',
  moderate: 'moderate',
  loud: 'loud',
} as const;

export interface Venue {
  id: number;
  name: string;
  category: VenueCategory;
  city: string;
  market: string;
  region: string;
  country: string;
  latitude: number;
  longitude: number;
  address: string;
  rating: number;
  /** 0-100 crowd pressure score */
  crowdScore: number;
  crowdLevel: VenueCrowdLevel;
  waitTimeMinutes: number;
  headcount: number;
  lineTrend: VenueLineTrend;
  seatingOdds: string;
  noiseLevel: VenueNoiseLevel;
  coverCost: string;
  bestTimeWindow: string;
  bestFor: string[];
  operatorGapNote: string;
  peakPressureWindow: string;
  reservationSignal: string;
  staffingSignal: string;
  dataSignalsTracked: string[];
  arrivalTips: string[];
  /** @nullable */
  sourceLabel?: string | null;
  /** @nullable */
  sourceUrl?: string | null;
  mapsUrl?: string;
  isWatchlisted: boolean;
  updatedAt: string;
}

export type VenueUpdateCategory = typeof VenueUpdateCategory[keyof typeof VenueUpdateCategory];


export const VenueUpdateCategory = {
  bar: 'bar',
  restaurant: 'restaurant',
  retail: 'retail',
  cafe: 'cafe',
  experience: 'experience',
} as const;

export type VenueUpdateNoiseLevel = typeof VenueUpdateNoiseLevel[keyof typeof VenueUpdateNoiseLevel];


export const VenueUpdateNoiseLevel = {
  quiet: 'quiet',
  moderate: 'moderate',
  loud: 'loud',
} as const;

export interface VenueUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  address?: string;
  category?: VenueUpdateCategory;
  bestFor?: string[];
  /** @minLength 1 */
  coverCost?: string;
  noiseLevel?: VenueUpdateNoiseLevel;
}

export interface VenueClaim {
  id: number;
  venueId: number;
  operatorUserId: string;
  createdAt: string;
}

export interface Market {
  market: string;
  city: string;
  region: string;
  country: string;
  venueCount: number;
}

export interface HeroStats {
  totalVenues: number;
  marketsCovered: number;
  liveReportsToday: number;
  averageCrowdScore: number;
  packedNow: number;
  openNow: number;
}

export interface HotZones {
  hottestPin: Venue;
  fastestMove: Venue;
  mostOpen: Venue;
}

export interface MarketGap {
  id: string;
  title: string;
  description: string;
}

export interface Comment {
  id: number;
  venueId: number;
  /** @nullable */
  parentCommentId?: number | null;
  authorName: string;
  message: string;
  likes: number;
  dislikes: number;
  createdAt: string;
}

export interface CommentInput {
  /** @minLength 1 */
  authorName: string;
  /** @minLength 1 */
  message: string;
  /** @nullable */
  parentCommentId?: number | null;
}

export type LiveReportCrowdLevel = typeof LiveReportCrowdLevel[keyof typeof LiveReportCrowdLevel];


export const LiveReportCrowdLevel = {
  open: 'open',
  lively: 'lively',
  packed: 'packed',
} as const;

export interface LiveReport {
  id: number;
  venueId: number;
  reporterName: string;
  crowdLevel: LiveReportCrowdLevel;
  waitTimeMinutes: number;
  vibeNote: string;
  createdAt: string;
}

export interface WatchlistEntry {
  id: number;
  userId: string;
  venueId: number;
  createdAt: string;
}

export type MyReport = LiveReport & {
  venueName: string;
};

export type MyComment = Comment & {
  venueName: string;
};

export interface MyActivity {
  reports: MyReport[];
  comments: MyComment[];
}

export type LiveReportInputCrowdLevel = typeof LiveReportInputCrowdLevel[keyof typeof LiveReportInputCrowdLevel];


export const LiveReportInputCrowdLevel = {
  open: 'open',
  lively: 'lively',
  packed: 'packed',
} as const;

export interface LiveReportInput {
  /** @minLength 1 */
  reporterName: string;
  crowdLevel: LiveReportInputCrowdLevel;
  /** @minimum 0 */
  waitTimeMinutes: number;
  /** @minLength 1 */
  vibeNote: string;
}

export type ListVenuesParams = {
market?: string;
category?: string;
search?: string;
/**
 * Sort order for results
 */
sort?: ListVenuesSort;
/**
 * Quick-pick user intent tag to filter by
 */
intent?: ListVenuesIntent;
};

export type ListVenuesSort = typeof ListVenuesSort[keyof typeof ListVenuesSort];


export const ListVenuesSort = {
  crowdScore: 'crowdScore',
  waitTime: 'waitTime',
  rating: 'rating',
  name: 'name',
  updated: 'updated',
} as const;

export type ListVenuesIntent = typeof ListVenuesIntent[keyof typeof ListVenuesIntent];


export const ListVenuesIntent = {
  dateNight: 'dateNight',
  noWait: 'noWait',
  retailDrops: 'retailDrops',
  liveMusic: 'liveMusic',
  patioEnergy: 'patioEnergy',
  lateNightFood: 'lateNightFood',
  speakeasy: 'speakeasy',
  interactiveBars: 'interactiveBars',
} as const;

export type GetHotZonesParams = {
market?: string;
};