# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ScenePulse API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues:
    get:
      operationId: listVenues
      tags: [venues]
      summary: List venues
      description: List venues with optional filtering, search, and sorting
      parameters:
        - name: market
          in: query
          schema: { type: string }
        - name: category
          in: query
          schema: { type: string }
        - name: search
          in: query
          schema: { type: string }
        - name: sort
          in: query
          description: Sort order for results
          schema:
            type: string
            enum: [crowdScore, waitTime, rating, name, updated]
        - name: intent
          in: query
          description: Quick-pick user intent tag to filter by
          schema:
            type: string
            enum: [dateNight, noWait, retailDrops, liveMusic, patioEnergy, lateNightFood, speakeasy, interactiveBars]
      responses:
        "200":
          description: List of venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
  /venues/{id}:
    get:
      operationId: getVenue
      tags: [venues]
      summary: Get a venue
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Venue detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "404":
          description: Not found

    patch:
      operationId: updateVenue
      tags: [venues]
      summary: Update a claimed venue's details (operator only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VenueUpdate"
      responses:
        "200":
          description: Updated venue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
        "401":
          description: Not authenticated
        "403":
          description: Venue not claimed by this operator
        "404":
          description: Not found
  /venues/{id}/claim:
    post:
      operationId: claimVenue
      tags: [venues]
      summary: Claim a venue as its operator
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "201":
          description: Claim created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueClaim"
        "401":
          description: Not authenticated
        "404":
          description: Venue not found
        "409":
          description: Venue already claimed by another operator
  /operator/venues:
    get:
      operationId: listOperatorVenues
      tags: [venues]
      summary: List venues claimed by the current operator
      responses:
        "200":
          description: Claimed venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
        "401":
          description: Not authenticated
...
  /venues/{venueId}/comments:
    get:
      operationId: listVenueComments
      tags: [comments]
      summary: List comments for a venue
      parameters:
        - name: venueId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: List of comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createVenueComment
      tags: [comments]
      summary: Post a comment on a venue's live scene
      parameters:
        - name: venueId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          description: Venue not found
  /venues/{venueId}/reports:
    get:
      operationId: listVenueReports
      tags: [reports]
      summary: List recent live reports for a venue
      parameters:
        - name: venueId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: List of live reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LiveReport"
    post:
      operationId: createVenueReport
      tags: [reports]
      summary: Submit a live crowd/wait/vibe report for a venue
      description: Updates the venue's live condition snapshot in addition to logging the report
      parameters:
        - name: venueId
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LiveReportInput"
      responses:
        "201":
          description: Created live report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LiveReport"
        "404":
          description: Venue not found
  /venues/{venueId}/comments/{commentId}/like:
    post:
      operationId: likeComment
      tags: [comments]
      summary: Like a comment
      parameters:
        - name: venueId
          in: path
          required: true
          schema: { type: integer }
        - name: commentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          description: Not found
  /venues/{venueId}/comments/{commentId}/dislike:
    post:
      operationId: dislikeComment
      tags: [comments]
      summary: Dislike a comment
      parameters:
        - name: venueId
          in: path
          required: true
          schema: { type: integer }
        - name: commentId
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"
        "404":
          description: Not found
components:
  schemas:
...
    Venue:
      type: object
      required:
        - id
        - name
        - category
        - city
        - market
        - region
        - country
        - latitude
        - longitude
        - address
        - rating
        - crowdScore
        - crowdLevel
        - waitTimeMinutes
        - headcount
        - lineTrend
        - seatingOdds
        - noiseLevel
        - coverCost
        - bestTimeWindow
        - bestFor
        - operatorGapNote
        - peakPressureWindow
        - reservationSignal
        - staffingSignal
        - dataSignalsTracked
        - arrivalTips
        - isWatchlisted
        - updatedAt
      properties:
        id: { type: integer }
        name: { type: string }
        category:
          type: string
          enum: [bar, restaurant, retail, cafe, experience]
        city: { type: string }
        market: { type: string }
        region: { type: string }
        country: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        address: { type: string }
        rating: { type: number }
        crowdScore:
          type: integer
          description: 0-100 crowd pressure score
        crowdLevel:
          type: string
          enum: [open, lively, packed]
        waitTimeMinutes: { type: integer }
        headcount: { type: integer }
        lineTrend:
          type: string
          enum: [rising, falling, steady]
        seatingOdds: { type: string }
        noiseLevel:
          type: string
          enum: [quiet, moderate, loud]
        coverCost: { type: string }
        bestTimeWindow: { type: string }
        bestFor:
          type: array
          items: { type: string }
        operatorGapNote: { type: string }
        peakPressureWindow: { type: string }
        reservationSignal: { type: string }
        staffingSignal: { type: string }
        dataSignalsTracked:
          type: array
          items: { type: string }
        arrivalTips:
          type: array
          items: { type: string }
        sourceLabel: { type: ["string", "null"] }
        sourceUrl: { type: ["string", "null"] }
        mapsUrl: { type: string }
        isWatchlisted: { type: boolean }
        updatedAt: { type: string, format: date-time }
    VenueUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        address: { type: string, minLength: 1 }
        category:
          type: string
          enum: [bar, restaurant, retail, cafe, experience]
        bestFor:
          type: array
          items: { type: string }
        coverCost: { type: string, minLength: 1 }
        noiseLevel:
          type: string
          enum: [quiet, moderate, loud]
    VenueClaim:
      type: object
      required: [id, venueId, operatorUserId, createdAt]
      properties:
        id: { type: integer }
        venueId: { type: integer }
        operatorUserId: { type: string }
        createdAt: { type: string, format: date-time }
...
    Comment:
      type: object
      required: [id, venueId, authorName, message, likes, dislikes, createdAt]
      properties:
        id: { type: integer }
        venueId: { type: integer }
        parentCommentId: { type: [integer, "null"] }
        authorName: { type: string }
        message: { type: string }
        likes: { type: integer }
        dislikes: { type: integer }
        createdAt: { type: string, format: date-time }
    CommentInput:
      type: object
      required: [authorName, message]
      properties:
        authorName: { type: string, minLength: 1 }
        message: { type: string, minLength: 1 }
        parentCommentId: { type: [integer, "null"] }
    LiveReport:
      type: object
      required:
        - id
        - venueId
        - reporterName
        - crowdLevel
        - waitTimeMinutes
        - vibeNote
        - createdAt
      properties:
        id: { type: integer }
        venueId: { type: integer }
        reporterName: { type: string }
        crowdLevel:
          type: string
          enum: [open, lively, packed]
        waitTimeMinutes: { type: integer }
        vibeNote: { type: string }
        createdAt: { type: string, format: date-time }
...
    LiveReportInput:
      type: object
      required: [reporterName, crowdLevel, waitTimeMinutes, vibeNote]
      properties:
        reporterName: { type: string, minLength: 1 }
        crowdLevel:
          type: string
          enum: [open, lively, packed]
        waitTimeMinutes: { type: integer, minimum: 0 }
        vibeNote: { type: string, minLength: 1 }